requirements.txt
aiopg==1.1.0
async-property==0.2.1
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
dataclasses==0.8
google-api-core==1.26.0
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
grpcio-tools==1.35.0
grpclib==0.4.1
h2==4.0.0
hpack==4.0.0
httplib2==0.19.0
hyperframe==6.0.0
idna==2.10
idna-ssl==1.1.0
multidict==5.1.0
numpy==1.19.5
packaging==20.9
peewee==3.14.1
peewee-async==0.7.1
protobuf==3.15.2
protobuf3-to-dict==0.1.5
protoc-gen-swagger==0.1.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
